Create a service for calculus for simulations
